move claims to pkg
