+++ components/director/internal/authenticator/claims/nsadapterClaimsValidator.go
@@ -2,6 +2,7 @@ package claims
 
 import (
 	"context"
+	"github.com/kyma-incubator/compass/components/director/pkg/token_claims"
 
 	"github.com/kyma-incubator/compass/components/hydrator/pkg/tenantmapping"
 	"github.com/pkg/errors"
@@ -15,7 +16,7 @@ func NewClaimsValidator() *claimsValidator {
 }
 
 // Validate validates the claims and asserts that the consumerTenant and externalTenant are not empty
-func (v *claimsValidator) Validate(_ context.Context, claims Claims) error {
+func (v *claimsValidator) Validate(_ context.Context, claims token_claims.Claims) error {
 	if err := claims.Valid(); err != nil {
 		return errors.Wrapf(err, "while validating claims")
 	}
