+++ components/director/internal/authenticator/claims/validators_test.go
@@ -3,6 +3,7 @@ package claims_test
 import (
 	"context"
 	"fmt"
+	"github.com/kyma-incubator/compass/components/director/pkg/token_claims"
 	"testing"
 
 	"github.com/kyma-incubator/compass/components/director/internal/authenticator/claims"
@@ -97,7 +98,7 @@ func TestValidator_Validate(t *testing.T) {
 		ApplicationServiceFn         func() *automock.ApplicationService
 		RuntimeCtxSvcFn              func() *automock.RuntimeCtxService
 		IntegrationSystemServiceFn   func() *automock.IntegrationSystemService
-		Claims                       claims.Claims
+		Claims                       token_claims.Claims
 		PersistenceFn                func() *persistenceautomock.PersistenceTx
 		TransactionerFn              func(persistTx *persistenceautomock.PersistenceTx) *persistenceautomock.Transactioner
 		ExpectedErr                  string
@@ -122,7 +123,7 @@ func TestValidator_Validate(t *testing.T) {
 		},
 		{
 			Name: "Fails when inner validation fails",
-			Claims: claims.Claims{
+			Claims: token_claims.Claims{
 				Tenant: map[string]string{
 					"consumerTenant": consumerTenantID,
 					"externalTenant": consumerExtTenantID,
@@ -459,16 +460,16 @@ func TestScopesValidator_Validate(t *testing.T) {
 	})
 }
 
-func getClaimsForRuntimeConsumerProviderFlow(consumerTenant, consumerExternalTenant, providerTenant, providerExtTenant, scopes, region, clientID string) claims.Claims {
+func getClaimsForRuntimeConsumerProviderFlow(consumerTenant, consumerExternalTenant, providerTenant, providerExtTenant, scopes, region, clientID string) token_claims.Claims {
 	return getClaimsForConsumerProviderFlow(consumer.Runtime, consumerTenant, consumerExternalTenant, consumerID, providerTenant, providerExtTenant, scopes, region, clientID)
 }
 
-func getClaimsForIntegrationSystemConsumerProviderFlow(consumerTenant, consumerExternalTenant, consumerID, providerTenant, providerExtTenant, scopes, region, clientID string) claims.Claims {
+func getClaimsForIntegrationSystemConsumerProviderFlow(consumerTenant, consumerExternalTenant, consumerID, providerTenant, providerExtTenant, scopes, region, clientID string) token_claims.Claims {
 	return getClaimsForConsumerProviderFlow(consumer.IntegrationSystem, consumerTenant, consumerExternalTenant, consumerID, providerTenant, providerExtTenant, scopes, region, clientID)
 }
 
-func getClaimsForConsumerProviderFlow(consumerType consumer.ConsumerType, consumerTenant, consumerExternalTenant, consumerID, providerTenant, providerExtTenant, scopes, region, clientID string) claims.Claims {
-	return claims.Claims{
+func getClaimsForConsumerProviderFlow(consumerType consumer.ConsumerType, consumerTenant, consumerExternalTenant, consumerID, providerTenant, providerExtTenant, scopes, region, clientID string) token_claims.Claims {
+	return token_claims.Claims{
 		Tenant: map[string]string{
 			tenantmapping.ConsumerTenantKey:         consumerTenant,
 			tenantmapping.ExternalTenantKey:         consumerExternalTenant,
@@ -484,8 +485,8 @@ func getClaimsForConsumerProviderFlow(consumerType consumer.ConsumerType, consum
 	}
 }
 
-func getClaims(intTenantID, extTenantID, scopes string) claims.Claims {
-	return claims.Claims{
+func getClaims(intTenantID, extTenantID, scopes string) token_claims.Claims {
+	return token_claims.Claims{
 		Tenant: map[string]string{
 			tenantmapping.ConsumerTenantKey: intTenantID,
 			tenantmapping.ExternalTenantKey: extTenantID,
