+++ components/director/pkg/auth-middleware/automock/claims_validator.go
@@ -5,8 +5,7 @@ package automock
 import (
 	context "context"
 
-	claims "github.com/kyma-incubator/compass/components/director/internal/authenticator/claims"
-
+	token_claims "github.com/kyma-incubator/compass/components/director/pkg/token_claims"
 	mock "github.com/stretchr/testify/mock"
 )
 
@@ -16,11 +15,11 @@ type ClaimsValidator struct {
 }
 
 // Validate provides a mock function with given fields: _a0, _a1
-func (_m *ClaimsValidator) Validate(_a0 context.Context, _a1 claims.Claims) error {
+func (_m *ClaimsValidator) Validate(_a0 context.Context, _a1 token_claims.Claims) error {
 	ret := _m.Called(_a0, _a1)
 
 	var r0 error
-	if rf, ok := ret.Get(0).(func(context.Context, claims.Claims) error); ok {
+	if rf, ok := ret.Get(0).(func(context.Context, token_claims.Claims) error); ok {
 		r0 = rf(_a0, _a1)
 	} else {
 		r0 = ret.Error(0)
