+++ components/director/pkg/auth-middleware/testdata/invalid.json
