+++ components/director/pkg/auth-middleware/testdata/jwks-public3.json
